third_party/blink/web_tests/wpt_internal/webxr/xrDevice_supportsSession_immersive_unsupported.https.html

<!DOCTYPE html>
<body>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/webxr/resources/webxr_util.js"></script>
  <script src="/webxr/resources/webxr_test_constants.js"></script>
  <script>
    // This test is chromium specific as though supportsSession is removed from
    // the spec it is still implemented, albeit as a deprecated method, in
    // chromium. Keeping this test until such time as it is removed.
    xr_promise_test(
      "supportsSession rejects when options not supported",
      (t) => {
      return navigator.xr.test.simulateDeviceConnection(VALID_NON_IMMERSIVE_DEVICE)
        .then( (controller) => {
          return promise_rejects_dom(
            t,
            "NotSupportedError",
            navigator.xr.supportsSession('immersive-vr')
          );
        });
    });
  </script>
</body>
